- Moved CustomInsertionPoints in their own package
