+++ src/burp/model/scanning/ScanItem.java
@@ -4,8 +4,8 @@ import burp.IHttpRequestResponse;
 import burp.IParameter;
 import burp.IResponseInfo;
 import burp.IScannerInsertionPoint;
-import burp.albinowaxUtils.ParamInsertionPoint;
-import burp.albinowaxUtils.RawInsertionPoint;
+import burp.model.insertionPoints.ParamInsertionPoint;
+import burp.model.insertionPoints.RawInsertionPoint;
 import burp.albinowaxUtils.Req;
 import burp.model.utilities.Utilities;
 import burp.view.ConfigurableSettings;
